requirements.txt
pySerial==3.5